refactor: Article 페이지에서useArticleCategory 훅 분리
